activitygen: input population and workposition (in streets) are float (not int). It's a number of people (or work positions) per length meter.
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@9021 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/activitygen/AGActivityGenHandler.cpp
@@ -165,9 +165,9 @@ AGActivityGenHandler::parseStreets(const SUMOSAXAttributes &attrs)
 	int work = 0;
 
 	if(attrs.hasAttribute(AGEN_ATTR_POPULATION))
-		pop = attrs.getInt(AGEN_ATTR_POPULATION);
+		pop = attrs.getFloat(AGEN_ATTR_POPULATION);
 	if(attrs.hasAttribute(AGEN_ATTR_OUT_WORKPOSITION))
-		work = attrs.getInt(AGEN_ATTR_OUT_WORKPOSITION);
+		work = attrs.getFloat(AGEN_ATTR_OUT_WORKPOSITION);
 
 	AGStreet str(attrs.getString(SUMO_ATTR_EDGE), net, pop, work);
 	myCity.streets.push_back(str);
